+++ cryptocurrency_parser/src/infrastructure/models/currency.py
@@ -13,6 +13,6 @@ class Currency(Base):
     id: Mapped[int] = mapped_column(primary_key=True, autoincrement=True)
     ticker: Mapped[str] = mapped_column(String(4))
     full_name: Mapped[str] = mapped_column(String(20), unique=True)
-    max_supply: Mapped[Optional[float]]
-    circulating_supply: Mapped[float]
+    max_supply: Mapped[Optional[int]]
+    circulating_supply: Mapped[int]
     last_updated: Mapped[Optional[datetime]]
